+++ lib/commands/app-management.js
@@ -334,27 +334,6 @@ export async function background(seconds) {
 
 /**
  * @this {import('../driver').AndroidDriver}
- * @returns {Promise<void>}
- */
-export async function startAUT() {
-  await this.adb.startApp({
-    pkg: /** @type {string} */ (this.opts.appPackage),
-    activity: this.opts.appActivity,
-    action: this.opts.intentAction,
-    category: this.opts.intentCategory,
-    flags: this.opts.intentFlags,
-    waitPkg: this.opts.appWaitPackage,
-    waitActivity: this.opts.appWaitActivity,
-    waitForLaunch: this.opts.appWaitForLaunch,
-    waitDuration: this.opts.appWaitDuration,
-    optionalIntentArguments: this.opts.optionalIntentArguments,
-    stopApp: !this.opts.dontStopAppOnReset,
-    user: this.opts.userProfile,
-  });
-}
-
-/**
- * @this {import('../driver').AndroidDriver}
  * @param {import('../driver').AndroidDriverOpts?} [opts=null]
  * @returns {Promise<void>}
  */
