+++ lib/commands/device/common.js
@@ -1,11 +1,18 @@
 import semver from 'semver';
 import _ from 'lodash';
 import { resetMockLocation, setMockLocationApp } from '../geolocation';
-import { pushSettingsApp } from '../settings-app';
 import { SETTINGS_HELPER_ID } from 'io.appium.settings';
 import { ensureDeviceLocale } from '../resources';
-import { hideKeyboardCompletely, initUnicodeKeyboard } from '../keyboard';
-import { createBaseADB, prepareEmulator, validatePackageActivityNames } from './utils';
+import {
+  hideKeyboardCompletely,
+  initUnicodeKeyboard,
+} from '../keyboard';
+import {
+  createBaseADB,
+  prepareEmulator,
+  validatePackageActivityNames,
+  pushSettingsApp,
+} from './utils';
 
 /**
  * @this {AndroidDriver}
