+++ lib/commands/file-actions.js
@@ -10,72 +10,6 @@ const CONTAINER_PATH_PATTERN = new RegExp(`^${CONTAINER_PATH_MARKER}([^/]+)/(.+)
 const ANDROID_MEDIA_RESCAN_INTENT = 'android.intent.action.MEDIA_SCANNER_SCAN_FILE';
 
 /**
- * Parses the actual destination path from the given value
- *
- * @param {string} remotePath The preformatted remote path, which looks like
- * `@my.app.id/my/path`
- * @returns {Array<string>} An array, where the first item is the parsed package
- * identifier and the second one is the actual destination path inside the package.
- * @throws {Error} If the given string cannot be parsed
- */
-function parseContainerPath(remotePath) {
-  const match = CONTAINER_PATH_PATTERN.exec(remotePath);
-  if (!match) {
-    throw new Error(
-      `It is expected that package identifier is separated from the relative path with a single slash. ` +
-        `'${remotePath}' is given instead`
-    );
-  }
-  return [match[1], path.posix.resolve(`/data/data/${match[1]}`, match[2])];
-}
-
-/**
- * Scans the given file/folder on the remote device
- * and adds matching items to the device's media library.
- * Exceptions are ignored and written into the log.
- *
- * @this {import('../driver').AndroidDriver}
- * @param {string} remotePath The file/folder path on the remote device
- */
-async function scanMedia(remotePath) {
-  this.log.debug(`Performing media scan of '${remotePath}'`);
-  try {
-    // https://github.com/appium/appium/issues/16184
-    if ((await this.adb.getApiLevel()) >= 29) {
-      await this.settingsApp.scanMedia(remotePath);
-    } else {
-      await this.adb.shell([
-        'am',
-        'broadcast',
-        '-a',
-        ANDROID_MEDIA_RESCAN_INTENT,
-        '-d',
-        `file://${remotePath}`,
-      ]);
-    }
-  } catch (e) {
-    const err = /** @type {any} */ (e);
-    // FIXME: what has a `stderr` prop?
-    this.log.warn(
-      `Ignoring an unexpected error upon media scanning of '${remotePath}': ${
-        err.stderr ?? err.message
-      }`
-    );
-  }
-}
-
-/**
- * A small helper, which escapes single quotes in paths,
- * so they are safe to be passed as arguments of shell commands
- *
- * @param {string} p The initial remote path
- * @returns {string} The escaped path value
- */
-function escapePath(p) {
-  return p.replace(/'/g, `\\'`);
-}
-
-/**
  * @this {import('../driver').AndroidDriver}
  * @param {string} remotePath
  * @returns {Promise<string>}
@@ -258,48 +192,6 @@ export async function mobileDeleteFile(opts) {
 }
 
 /**
- * Factory providing filesystem test functions using ADB
- * @param {ADB} adb
- */
-function createFSTests(adb) {
-  /**
-   *
-   * @param {string} p
-   * @param {'d'|'f'|'e'} op
-   * @param {string} [runAs]
-   * @returns
-   */
-  const performRemoteFsCheck = async (p, op, runAs) => {
-    const passFlag = '__PASS__';
-    const checkCmd = `[ -${op} '${escapePath(p)}' ] && echo ${passFlag}`;
-    const fullCmd = runAs ? `run-as ${runAs} ${checkCmd}` : checkCmd;
-    try {
-      return _.includes(await adb.shell([fullCmd]), passFlag);
-    } catch (ign) {
-      return false;
-    }
-  };
-
-  /**
-   * @param {string} p
-   * @param {string} [runAs]
-   */
-  const isFile = async (p, runAs) => await performRemoteFsCheck(p, 'f', runAs);
-  /**
-   * @param {string} p
-   * @param {string} [runAs]
-   */
-  const isDir = async (p, runAs) => await performRemoteFsCheck(p, 'd', runAs);
-  /**
-   * @param {string} p
-   * @param {string} [runAs]
-   */
-  const isPresent = async (p, runAs) => await performRemoteFsCheck(p, 'e', runAs);
-
-  return {isFile, isDir, isPresent};
-}
-
-/**
  * Deletes the given folder or file from the remote device
  *
  * @param {ADB} adb
@@ -360,6 +252,119 @@ async function deleteFileOrFolder(adb, remotePath) {
   }
   return true;
 }
+
+// #region Internal helpers
+
+
+/**
+ * Parses the actual destination path from the given value
+ *
+ * @param {string} remotePath The preformatted remote path, which looks like
+ * `@my.app.id/my/path`
+ * @returns {Array<string>} An array, where the first item is the parsed package
+ * identifier and the second one is the actual destination path inside the package.
+ * @throws {Error} If the given string cannot be parsed
+ */
+function parseContainerPath(remotePath) {
+  const match = CONTAINER_PATH_PATTERN.exec(remotePath);
+  if (!match) {
+    throw new Error(
+      `It is expected that package identifier is separated from the relative path with a single slash. ` +
+        `'${remotePath}' is given instead`
+    );
+  }
+  return [match[1], path.posix.resolve(`/data/data/${match[1]}`, match[2])];
+}
+
+/**
+ * Scans the given file/folder on the remote device
+ * and adds matching items to the device's media library.
+ * Exceptions are ignored and written into the log.
+ *
+ * @this {import('../driver').AndroidDriver}
+ * @param {string} remotePath The file/folder path on the remote device
+ */
+async function scanMedia(remotePath) {
+  this.log.debug(`Performing media scan of '${remotePath}'`);
+  try {
+    // https://github.com/appium/appium/issues/16184
+    if ((await this.adb.getApiLevel()) >= 29) {
+      await this.settingsApp.scanMedia(remotePath);
+    } else {
+      await this.adb.shell([
+        'am',
+        'broadcast',
+        '-a',
+        ANDROID_MEDIA_RESCAN_INTENT,
+        '-d',
+        `file://${remotePath}`,
+      ]);
+    }
+  } catch (e) {
+    const err = /** @type {any} */ (e);
+    // FIXME: what has a `stderr` prop?
+    this.log.warn(
+      `Ignoring an unexpected error upon media scanning of '${remotePath}': ${
+        err.stderr ?? err.message
+      }`
+    );
+  }
+}
+
+/**
+ * A small helper, which escapes single quotes in paths,
+ * so they are safe to be passed as arguments of shell commands
+ *
+ * @param {string} p The initial remote path
+ * @returns {string} The escaped path value
+ */
+function escapePath(p) {
+  return p.replace(/'/g, `\\'`);
+}
+
+/**
+ * Factory providing filesystem test functions using ADB
+ * @param {ADB} adb
+ */
+function createFSTests(adb) {
+  /**
+   *
+   * @param {string} p
+   * @param {'d'|'f'|'e'} op
+   * @param {string} [runAs]
+   * @returns
+   */
+  const performRemoteFsCheck = async (p, op, runAs) => {
+    const passFlag = '__PASS__';
+    const checkCmd = `[ -${op} '${escapePath(p)}' ] && echo ${passFlag}`;
+    const fullCmd = runAs ? `run-as ${runAs} ${checkCmd}` : checkCmd;
+    try {
+      return _.includes(await adb.shell([fullCmd]), passFlag);
+    } catch (ign) {
+      return false;
+    }
+  };
+
+  /**
+   * @param {string} p
+   * @param {string} [runAs]
+   */
+  const isFile = async (p, runAs) => await performRemoteFsCheck(p, 'f', runAs);
+  /**
+   * @param {string} p
+   * @param {string} [runAs]
+   */
+  const isDir = async (p, runAs) => await performRemoteFsCheck(p, 'd', runAs);
+  /**
+   * @param {string} p
+   * @param {string} [runAs]
+   */
+  const isPresent = async (p, runAs) => await performRemoteFsCheck(p, 'e', runAs);
+
+  return {isFile, isDir, isPresent};
+}
+
+// #endregion
 
 /**
  * @typedef {import('appium-adb').ADB} ADB
