+++ lib/commands/intent.js
@@ -24,80 +24,6 @@ const SUPPORTED_EXTRA_TYPES = [
 const API_LEVEL_ANDROID_8 = 26;
 
 /**
- *
- * @param {import('./types').IntentOpts} opts
- * @returns {string[]}
- */
-function parseIntentSpec(opts = {}) {
-  const {intent, action, uri, mimeType, identifier, categories, component, extras, flags} = opts;
-  const resultArgs = [];
-  if (intent) {
-    resultArgs.push(intent);
-  }
-  if (action) {
-    resultArgs.push('-a', action);
-  }
-  if (uri) {
-    resultArgs.push('-d', uri);
-  }
-  if (mimeType) {
-    resultArgs.push('-t', mimeType);
-  }
-  if (!_.isNil(identifier)) {
-    resultArgs.push('-i', identifier);
-  }
-  if (categories) {
-    if (_.isArray(categories)) {
-      resultArgs.push(..._.flatMap(categories.map((cName) => ['-c', cName])));
-    } else {
-      resultArgs.push('-c', categories);
-    }
-  }
-  if (component) {
-    resultArgs.push('-n', component);
-  }
-  if (opts.package) {
-    resultArgs.push('-p', opts.package);
-  }
-  if (extras) {
-    if (!_.isArray(extras)) {
-      throw new errors.InvalidArgumentError(`'extras' must be an array`);
-    }
-    for (const item of extras) {
-      if (!_.isArray(item)) {
-        throw new errors.InvalidArgumentError(`Extra argument '${item}' must be an array`);
-      }
-      const [type, key, value] = item;
-      if (!_.includes(SUPPORTED_EXTRA_TYPES, type)) {
-        throw new errors.InvalidArgumentError(
-          `Extra argument type '${type}' is not known. ` +
-            `Supported intent argument types are: ${SUPPORTED_EXTRA_TYPES}`
-        );
-      }
-      if (_.isEmpty(key) || (_.isString(key) && _.trim(key) === '')) {
-        throw new errors.InvalidArgumentError(
-          `Extra argument's key in '${JSON.stringify(item)}' must be a valid string identifier`
-        );
-      }
-      if (type === NO_VALUE_ARG_TYPE) {
-        resultArgs.push(`--e${type}`, key);
-      } else if (_.isUndefined(value)) {
-        throw new errors.InvalidArgumentError(
-          `Intent argument type '${type}' in '${JSON.stringify(item)}' requires a ` +
-            `valid value to be provided`
-        );
-      } else {
-        resultArgs.push(`--e${type}`, key, value);
-      }
-    }
-  }
-  if (flags) {
-    resultArgs.push('-f', flags);
-  }
-  return resultArgs;
-}
-
-/**
  * @deprecated
  * @this {import('../driver').AndroidDriver}
  * @param {string} appPackage
@@ -242,6 +168,84 @@ export async function mobileStopService(opts = {}) {
   cmd.push(...parseIntentSpec(opts));
   return await this.adb.shell(cmd);
 }
+
+// #region Internal helpers
+
+/**
+ *
+ * @param {import('./types').IntentOpts} opts
+ * @returns {string[]}
+ */
+function parseIntentSpec(opts = {}) {
+  const {intent, action, uri, mimeType, identifier, categories, component, extras, flags} = opts;
+  const resultArgs = [];
+  if (intent) {
+    resultArgs.push(intent);
+  }
+  if (action) {
+    resultArgs.push('-a', action);
+  }
+  if (uri) {
+    resultArgs.push('-d', uri);
+  }
+  if (mimeType) {
+    resultArgs.push('-t', mimeType);
+  }
+  if (!_.isNil(identifier)) {
+    resultArgs.push('-i', identifier);
+  }
+  if (categories) {
+    if (_.isArray(categories)) {
+      resultArgs.push(..._.flatMap(categories.map((cName) => ['-c', cName])));
+    } else {
+      resultArgs.push('-c', categories);
+    }
+  }
+  if (component) {
+    resultArgs.push('-n', component);
+  }
+  if (opts.package) {
+    resultArgs.push('-p', opts.package);
+  }
+  if (extras) {
+    if (!_.isArray(extras)) {
+      throw new errors.InvalidArgumentError(`'extras' must be an array`);
+    }
+    for (const item of extras) {
+      if (!_.isArray(item)) {
+        throw new errors.InvalidArgumentError(`Extra argument '${item}' must be an array`);
+      }
+      const [type, key, value] = item;
+      if (!_.includes(SUPPORTED_EXTRA_TYPES, type)) {
+        throw new errors.InvalidArgumentError(
+          `Extra argument type '${type}' is not known. ` +
+            `Supported intent argument types are: ${SUPPORTED_EXTRA_TYPES}`
+        );
+      }
+      if (_.isEmpty(key) || (_.isString(key) && _.trim(key) === '')) {
+        throw new errors.InvalidArgumentError(
+          `Extra argument's key in '${JSON.stringify(item)}' must be a valid string identifier`
+        );
+      }
+      if (type === NO_VALUE_ARG_TYPE) {
+        resultArgs.push(`--e${type}`, key);
+      } else if (_.isUndefined(value)) {
+        throw new errors.InvalidArgumentError(
+          `Intent argument type '${type}' in '${JSON.stringify(item)}' requires a ` +
+            `valid value to be provided`
+        );
+      } else {
+        resultArgs.push(`--e${type}`, key, value);
+      }
+    }
+  }
+  if (flags) {
+    resultArgs.push('-f', flags);
+  }
+  return resultArgs;
+}
+
+// #endregion
 
 /**
  * @typedef {import('appium-adb').ADB} ADB
