+++ lib/commands/keyboard.js
@@ -81,7 +81,7 @@ export async function mobilePerformEditorAction(opts) {
   await this.settingsApp.performEditorAction(action);
 }
 
-// #region Internal Methods
+// #region Internal Helpers
 
 /**
  * @deprecated
