+++ lib/commands/lock.js
@@ -85,7 +85,7 @@ export async function mobileUnlock(opts = {}) {
   }
 }
 
-// #region Internal methods
+// #region Internal helpers
 
 /**
  *
