+++ lib/commands/log.js
@@ -6,28 +6,6 @@ import log from '../logger';
 
 const GET_SERVER_LOGS_FEATURE = 'get_server_logs';
 
-/**
- * @param {string} sessionId
- * @returns {string}
- */
-const WEBSOCKET_ENDPOINT = (sessionId) =>
-  `${DEFAULT_WS_PATHNAME_PREFIX}/session/${sessionId}/appium/device/logcat`;
-
-/**
- *
- * @see {@link https://github.com/SeleniumHQ/selenium/blob/0d425676b3c9df261dd641917f867d4d5ce7774d/java/client/src/org/openqa/selenium/logging/LogEntry.java}
- * @param {number} timestamp
- * @param {string} level
- * @param {string} message
- */
-function toLogRecord(timestamp, level, message) {
-  return {
-    timestamp,
-    level,
-    message,
-  };
-}
-
 export const supportedLogTypes = {
   logcat: {
     description: 'Logs for Android applications on real device and emulators via ADB',
@@ -197,6 +175,32 @@ export async function getLog(logType) {
   // XXX why doesn't `super` work here?
   return await BaseDriver.prototype.getLog.call(this, logType);
 }
+
+// #region Internal helpers
+
+/**
+ * @param {string} sessionId
+ * @returns {string}
+ */
+const WEBSOCKET_ENDPOINT = (sessionId) =>
+  `${DEFAULT_WS_PATHNAME_PREFIX}/session/${sessionId}/appium/device/logcat`;
+
+/**
+ *
+ * @see {@link https://github.com/SeleniumHQ/selenium/blob/0d425676b3c9df261dd641917f867d4d5ce7774d/java/client/src/org/openqa/selenium/logging/LogEntry.java}
+ * @param {number} timestamp
+ * @param {string} level
+ * @param {string} message
+ */
+function toLogRecord(timestamp, level, message) {
+  return {
+    timestamp,
+    level,
+    message,
+  };
+}
+
+// #endregion
 
 /**
  * @typedef {import('appium-adb').ADB} ADB
