+++ lib/commands/media-projection.js
@@ -9,67 +9,6 @@ const MIN_API_LEVEL = 29;
 const DEFAULT_FILENAME_FORMAT = 'YYYY-MM-DDTHH-mm-ss';
 
 /**
- *
- * @param {string} localFile
- * @param {string} [remotePath]
- * @param {import('./types').StopMediaProjectionRecordingOpts} uploadOptions
- * @returns
- */
-async function uploadRecordedMedia(localFile, remotePath, uploadOptions = {}) {
-  if (_.isEmpty(remotePath)) {
-    return (await util.toInMemoryBase64(localFile)).toString();
-  }
-
-  const {
-    user,
-    pass,
-    method,
-    headers,
-    fileFieldName,
-    formFields,
-    uploadTimeout: timeout,
-  } = uploadOptions;
-  /**
-   * @type {Omit<import('./types').StopMediaProjectionRecordingOpts, 'uploadTimeout'> & {auth?: {user: string, pass: string}, timeout?: number}}
-   */
-  const options = {
-    method: method || 'PUT',
-    headers,
-    fileFieldName,
-    formFields,
-    timeout,
-  };
-  if (user && pass) {
-    options.auth = {user, pass};
-  }
-  await net.uploadFile(localFile, /** @type {string} */ (remotePath), options);
-  return '';
-}
-
-/**
- *
- * @param {string} name
- * @returns {string}
- */
-function adjustMediaExtension(name) {
-  return _.toLower(name).endsWith(DEFAULT_EXT) ? name : `${name}${DEFAULT_EXT}`;
-}
-
-/**
- *
- * @param {ADB} adb
- */
-async function verifyMediaProjectionRecordingIsSupported(adb) {
-  const apiLevel = await adb.getApiLevel();
-  if (apiLevel < MIN_API_LEVEL) {
-    throw new Error(
-      `Media projection-based recording is not available on API Level ${apiLevel}. ` +
-        `Minimum required API Level is ${MIN_API_LEVEL}.`
-    );
-  }
-}
-
-/**
  * @this {import('../driver').AndroidDriver}
  * @param {import('./types').StartMediaProjectionRecordingOpts} [options={}]
  * @returns {Promise<boolean>}
@@ -141,6 +80,71 @@ export async function mobileStopMediaProjectionRecording(options = {}) {
     await fs.rimraf(path.dirname(recentRecordingPath));
   }
 }
+
+// #region Internal helpers
+
+/**
+ *
+ * @param {string} localFile
+ * @param {string} [remotePath]
+ * @param {import('./types').StopMediaProjectionRecordingOpts} uploadOptions
+ * @returns
+ */
+async function uploadRecordedMedia(localFile, remotePath, uploadOptions = {}) {
+  if (_.isEmpty(remotePath)) {
+    return (await util.toInMemoryBase64(localFile)).toString();
+  }
+
+  const {
+    user,
+    pass,
+    method,
+    headers,
+    fileFieldName,
+    formFields,
+    uploadTimeout: timeout,
+  } = uploadOptions;
+  /**
+   * @type {Omit<import('./types').StopMediaProjectionRecordingOpts, 'uploadTimeout'> & {auth?: {user: string, pass: string}, timeout?: number}}
+   */
+  const options = {
+    method: method || 'PUT',
+    headers,
+    fileFieldName,
+    formFields,
+    timeout,
+  };
+  if (user && pass) {
+    options.auth = {user, pass};
+  }
+  await net.uploadFile(localFile, /** @type {string} */ (remotePath), options);
+  return '';
+}
+
+/**
+ *
+ * @param {string} name
+ * @returns {string}
+ */
+function adjustMediaExtension(name) {
+  return _.toLower(name).endsWith(DEFAULT_EXT) ? name : `${name}${DEFAULT_EXT}`;
+}
+
+/**
+ *
+ * @param {ADB} adb
+ */
+async function verifyMediaProjectionRecordingIsSupported(adb) {
+  const apiLevel = await adb.getApiLevel();
+  if (apiLevel < MIN_API_LEVEL) {
+    throw new Error(
+      `Media projection-based recording is not available on API Level ${apiLevel}. ` +
+        `Minimum required API Level is ${MIN_API_LEVEL}.`
+    );
+  }
+}
+
+// #endregion
 
 /**
  * @typedef {import('appium-adb').ADB} ADB
