+++ lib/commands/performance.js
@@ -57,6 +57,81 @@ const MEMINFO_TITLES = Object.freeze({
 const RETRY_PAUSE_MS = 1000;
 
 /**
+ * @this {AndroidDriver}
+ * @returns {Promise<import('./types').PerformanceDataType[]>}
+ */
+export async function getPerformanceDataTypes() {
+  return /** @type {import('./types').PerformanceDataType[]} */ (
+    _.keys(SUPPORTED_PERFORMANCE_DATA_TYPES)
+  );
+}
+
+/**
+ * @this {AndroidDriver}
+ * @param {string} packageName
+ * @param {string} dataType
+ * @param {number} [retries=2]
+ * @returns {Promise<any[][]>}
+ */
+export async function getPerformanceData(packageName, dataType, retries = 2) {
+  let result;
+  switch (_.toLower(dataType)) {
+    case 'batteryinfo':
+      result = await getBatteryInfo.call(this, retries);
+      break;
+    case 'cpuinfo':
+      result = await getCPUInfo.call(this, packageName, retries);
+      break;
+    case 'memoryinfo':
+      result = await getMemoryInfo.call(this, packageName, retries);
+      break;
+    case 'networkinfo':
+      result = await getNetworkTrafficInfo.call(this, retries);
+      break;
+    default:
+      throw new Error(
+        `No performance data of type '${dataType}' found. ` +
+          `Only the following values are supported: ${JSON.stringify(
+            SUPPORTED_PERFORMANCE_DATA_TYPES,
+            [' '],
+            2
+          )}`
+      );
+  }
+  return /** @type {any[][]} */ (result);
+}
+
+/**
+ * Retrieves performance data about the given Android subsystem.
+ * The data is parsed from the output of the dumpsys utility.
+ *
+ * The output depends on the selected subsystem.
+ * It is orginized into a table, where the first row represent column names
+ * and the following rows represent the sampled data for each column.
+ * Example output for different data types:
+ * - batteryinfo: [[power], [23]]
+ * - memory info: [[totalPrivateDirty, nativePrivateDirty, dalvikPrivateDirty, eglPrivateDirty, glPrivateDirty, totalPss,
+ *   nativePss, dalvikPss, eglPss, glPss, nativeHeapAllocatedSize, nativeHeapSize], [18360, 8296, 6132, null, null, 42588, 8406, 7024, null, null, 26519, 10344]]
+ * - networkinfo: [[bucketStart, activeTime, rxBytes, rxPackets, txBytes, txPackets, operations, bucketDuration,],
+ *   [1478091600000, null, 1099075, 610947, 928, 114362, 769, 0, 3600000], [1478095200000, null, 1306300, 405997, 509, 46359, 370, 0, 3600000]]
+ *
+ *   [[st, activeTime, rb, rp, tb, tp, op, bucketDuration], [1478088000, null, null, 32115296, 34291, 2956805, 25705, 0, 3600],
+ *   [1478091600, null, null, 2714683, 11821, 1420564, 12650, 0, 3600], [1478095200, null, null, 10079213, 19962, 2487705, 20015, 0, 3600],
+ *   [1478098800, null, null, 4444433, 10227, 1430356, 10493, 0, 3600]]
+ * - cpuinfo: [[user, kernel], [0.9, 1.3]]
+ *
+ * @this {AndroidDriver}
+ * @param {import('./types').PerformanceDataOpts} opts
+ * @returns {Promise<any[][]>}
+ */
+export async function mobileGetPerformanceData(opts) {
+  const {packageName, dataType} = requireArgs(['packageName', 'dataType'], opts);
+  return await this.getPerformanceData(packageName, dataType);
+}
+
+// #region Internal helpers
+
+/**
  * API level between 18 and 30
  * ['<System Type>', '<Memory Type>', <pss total>, <private dirty>, <private clean>, <swapPss dirty>, <heap size>, <heap alloc>, <heap free>]
  * except 'TOTAL', which skips the second type name
@@ -435,90 +510,7 @@ async function getBatteryInfo(retries = 2) {
   });
 }
 
-/**
- * @this {AndroidDriver}
- * @returns {Promise<import('./types').PerformanceDataType[]>}
- */
-export async function getPerformanceDataTypes() {
-  return /** @type {import('./types').PerformanceDataType[]} */ (
-    _.keys(SUPPORTED_PERFORMANCE_DATA_TYPES)
-  );
-}
-
-/**
- * @this {AndroidDriver}
- * @param {string} packageName
- * @param {string} dataType
- * @param {number} [retries=2]
- * @returns {Promise<any[][]>}
- */
-export async function getPerformanceData(packageName, dataType, retries = 2) {
-  let result;
-  switch (_.toLower(dataType)) {
-    case 'batteryinfo':
-      result = await getBatteryInfo.call(this, retries);
-      break;
-    case 'cpuinfo':
-      result = await getCPUInfo.call(this, packageName, retries);
-      break;
-    case 'memoryinfo':
-      result = await getMemoryInfo.call(this, packageName, retries);
-      break;
-    case 'networkinfo':
-      result = await getNetworkTrafficInfo.call(this, retries);
-      break;
-    default:
-      throw new Error(
-        `No performance data of type '${dataType}' found. ` +
-          `Only the following values are supported: ${JSON.stringify(
-            SUPPORTED_PERFORMANCE_DATA_TYPES,
-            [' '],
-            2
-          )}`
-      );
-  }
-  return /** @type {any[][]} */ (result);
-}
-
-/**
- * Retrieves performance data about the given Android subsystem.
- * The data is parsed from the output of the dumpsys utility.
- *
- * The output depends on the selected subsystem.
- * It is orginized into a table, where the first row represent column names
- * and the following rows represent the sampled data for each column.
- * Example output for different data types:
- * - batteryinfo: [[power], [23]]
- * - memory info: [[totalPrivateDirty, nativePrivateDirty, dalvikPrivateDirty, eglPrivateDirty, glPrivateDirty, totalPss,
- *   nativePss, dalvikPss, eglPss, glPss, nativeHeapAllocatedSize, nativeHeapSize], [18360, 8296, 6132, null, null, 42588, 8406, 7024, null, null, 26519, 10344]]
- * - networkinfo: [[bucketStart, activeTime, rxBytes, rxPackets, txBytes, txPackets, operations, bucketDuration,],
- *   [1478091600000, null, 1099075, 610947, 928, 114362, 769, 0, 3600000], [1478095200000, null, 1306300, 405997, 509, 46359, 370, 0, 3600000]]
- *
- *   [[st, activeTime, rb, rp, tb, tp, op, bucketDuration], [1478088000, null, null, 32115296, 34291, 2956805, 25705, 0, 3600],
- *   [1478091600, null, null, 2714683, 11821, 1420564, 12650, 0, 3600], [1478095200, null, null, 10079213, 19962, 2487705, 20015, 0, 3600],
- *   [1478098800, null, null, 4444433, 10227, 1430356, 10493, 0, 3600]]
- * - cpuinfo: [[user, kernel], [0.9, 1.3]]
- *
- * @this {AndroidDriver}
- * @param {import('./types').PerformanceDataOpts} opts
- * @returns {Promise<any[][]>}
- */
-export async function mobileGetPerformanceData(opts) {
-  const {packageName, dataType} = requireArgs(['packageName', 'dataType'], opts);
-  return await this.getPerformanceData(packageName, dataType);
-}
-
-export {
-  BATTERY_KEYS,
-  CPU_KEYS,
-  MEMORY_KEYS,
-  NETWORK_KEYS,
-  SUPPORTED_PERFORMANCE_DATA_TYPES,
-  getBatteryInfo,
-  getCPUInfo,
-  getMemoryInfo,
-  getNetworkTrafficInfo,
-};
+// #endregion
 
 /**
  * @typedef {import('../driver').AndroidDriver} AndroidDriver
