+++ lib/commands/permissions.js
@@ -25,70 +25,6 @@ const PERMISSIONS_TYPE = Object.freeze({
 });
 
 /**
- * @this {AndroidDriver}
- * @param {*} permissions
- * @param {*} appPackage
- * @param {*} action
- * @todo FIXME: type this
- */
-async function changePermissionsViaPm(permissions, appPackage, action) {
-  if (!_.values(PM_ACTION).includes(action)) {
-    throw new errors.InvalidArgumentError(
-      `Unknown action '${action}'. ` +
-        `Only ${JSON.stringify(_.values(PM_ACTION))} actions are supported`
-    );
-  }
-
-  let affectedPermissions = _.isArray(permissions) ? permissions : [permissions];
-  if (_.toLower(permissions) === ALL_PERMISSIONS_MAGIC) {
-    const dumpsys = await this.adb.shell(['dumpsys', 'package', appPackage]);
-    const grantedPermissions = await this.adb.getGrantedPermissions(appPackage, dumpsys);
-    if (action === PM_ACTION.GRANT) {
-      const reqPermissons = await this.adb.getReqPermissions(appPackage, dumpsys);
-      affectedPermissions = _.difference(reqPermissons, grantedPermissions);
-    } else {
-      affectedPermissions = grantedPermissions;
-    }
-    if (_.isEmpty(affectedPermissions)) {
-      this.log.info(`'${appPackage}' contains no permissions to ${action}`);
-      return;
-    }
-  }
-
-  if (action === PM_ACTION.GRANT) {
-    await this.adb.grantPermissions(appPackage, affectedPermissions);
-  } else {
-    await B.all(affectedPermissions.map((name) => this.adb.revokePermission(appPackage, name)));
-  }
-}
-/**
- * @this {AndroidDriver}
- * @param {string|string[]} permissions
- * @param {string} appPackage
- * @param {typeof APPOPS_ACTION[keyof APPOPS_ACTION]} action
- */
-async function changePermissionsViaAppops(permissions, appPackage, action) {
-  if (!_.values(APPOPS_ACTION).includes(action)) {
-    throw new errors.InvalidArgumentError(
-      `Unknown action '${action}'. ` +
-        `Only ${JSON.stringify(_.values(APPOPS_ACTION))} actions are supported`
-    );
-  }
-  if (_.isString(permissions) && _.toLower(permissions) === ALL_PERMISSIONS_MAGIC) {
-    throw new errors.InvalidArgumentError(
-      `'${ALL_PERMISSIONS_MAGIC}' permission is only supported for ` +
-        `'${PERMISSION_TARGET.PM}' target. ` +
-        `Check AppOpsManager.java from Android platform sources to get the full list of supported AppOps permissions`
-    );
-  }
-
-  const promises = (_.isArray(permissions) ? permissions : [permissions]).map((permission) =>
-    this.adb.shell(['appops', 'set', appPackage, permission, action])
-  );
-  await B.all(promises);
-}
-
-/**
  * @this {import('../driver').AndroidDriver}
  * @param {import('./types').ChangePermissionsOpts} opts
  * @returns {Promise<void>}
@@ -155,6 +91,73 @@ export async function mobileGetPermissions(opts = {}) {
   }
   return await actionFunc(/** @type {string} */ (appPackage));
 }
+
+// #region Internal helpers
+
+/**
+ * @this {AndroidDriver}
+ * @param {string|string[]} permissions
+ * @param {string} appPackage
+ * @param {import('type-fest').ValueOf<PM_ACTION>} action
+ */
+async function changePermissionsViaPm(permissions, appPackage, action) {
+  if (!_.values(PM_ACTION).includes(action)) {
+    throw new errors.InvalidArgumentError(
+      `Unknown action '${action}'. ` +
+        `Only ${JSON.stringify(_.values(PM_ACTION))} actions are supported`
+    );
+  }
+
+  let affectedPermissions = _.isArray(permissions) ? permissions : [permissions];
+  if (_.isString(permissions) && _.toLower(permissions) === ALL_PERMISSIONS_MAGIC) {
+    const dumpsys = await this.adb.shell(['dumpsys', 'package', appPackage]);
+    const grantedPermissions = await this.adb.getGrantedPermissions(appPackage, dumpsys);
+    if (action === PM_ACTION.GRANT) {
+      const reqPermissons = await this.adb.getReqPermissions(appPackage, dumpsys);
+      affectedPermissions = _.difference(reqPermissons, grantedPermissions);
+    } else {
+      affectedPermissions = grantedPermissions;
+    }
+    if (_.isEmpty(affectedPermissions)) {
+      this.log.info(`'${appPackage}' contains no permissions to ${action}`);
+      return;
+    }
+  }
+
+  if (action === PM_ACTION.GRANT) {
+    await this.adb.grantPermissions(appPackage, affectedPermissions);
+  } else {
+    await B.all(affectedPermissions.map((name) => this.adb.revokePermission(appPackage, name)));
+  }
+}
+/**
+ * @this {AndroidDriver}
+ * @param {string|string[]} permissions
+ * @param {string} appPackage
+ * @param {import('type-fest').ValueOf<APPOPS_ACTION>} action
+ */
+async function changePermissionsViaAppops(permissions, appPackage, action) {
+  if (!_.values(APPOPS_ACTION).includes(action)) {
+    throw new errors.InvalidArgumentError(
+      `Unknown action '${action}'. ` +
+        `Only ${JSON.stringify(_.values(APPOPS_ACTION))} actions are supported`
+    );
+  }
+  if (_.isString(permissions) && _.toLower(permissions) === ALL_PERMISSIONS_MAGIC) {
+    throw new errors.InvalidArgumentError(
+      `'${ALL_PERMISSIONS_MAGIC}' permission is only supported for ` +
+        `'${PERMISSION_TARGET.PM}' target. ` +
+        `Check AppOpsManager.java from Android platform sources to get the full list of supported AppOps permissions`
+    );
+  }
+
+  const promises = (_.isArray(permissions) ? permissions : [permissions]).map((permission) =>
+    this.adb.shell(['appops', 'set', appPackage, permission, action])
+  );
+  await B.all(promises);
+}
+
+// #endregion
 
 /**
  * @typedef {import('appium-adb').ADB} ADB
