+++ lib/commands/recordscreen.js
@@ -17,6 +17,155 @@ const FFMPEG_BINARY = `ffmpeg${system.isWindows() ? '.exe' : ''}`;
 
 /**
  *
+ * @this {import('../driver').AndroidDriver}
+ * @param {import('./types').StartScreenRecordingOpts} [options={}]
+ * @returns {Promise<string>}
+ */
+export async function startRecordingScreen(options = {}) {
+  await verifyScreenRecordIsSupported(this.adb, this.isEmulator());
+
+  let result = '';
+  const {
+    videoSize,
+    timeLimit = DEFAULT_RECORDING_TIME_SEC,
+    bugReport,
+    bitRate,
+    forceRestart,
+  } = options;
+  if (!forceRestart) {
+    result = await this.stopRecordingScreen(options);
+  }
+
+  if (await terminateBackgroundScreenRecording(this.adb, true)) {
+    this.log.warn(
+      `There were some ${SCREENRECORD_BINARY} process leftovers running ` +
+        `in the background. Make sure you stop screen recording each time after it is started, ` +
+        `otherwise the recorded media might quickly exceed all the free space on the device under test.`
+    );
+  }
+
+  if (!_.isEmpty(this._screenRecordingProperties)) {
+    // XXX: this doesn't need to be done in serial, does it?
+    for (const record of this._screenRecordingProperties.records || []) {
+      await this.adb.rimraf(record);
+    }
+    this._screenRecordingProperties = undefined;
+  }
+
+  const timeout = parseFloat(String(timeLimit));
+  if (isNaN(timeout) || timeout > MAX_TIME_SEC || timeout <= 0) {
+    throw new Error(
+      `The timeLimit value must be in range [1, ${MAX_TIME_SEC}] seconds. ` +
+        `The value of '${timeLimit}' has been passed instead.`
+    );
+  }
+
+  this._screenRecordingProperties = {
+    timer: new timing.Timer().start(),
+    videoSize,
+    timeLimit,
+    currentTimeLimit: timeLimit,
+    bitRate,
+    bugReport,
+    records: [],
+    recordingProcess: null,
+    stopped: false,
+  };
+  await scheduleScreenRecord(this.adb, this._screenRecordingProperties, this.log);
+  return result;
+}
+
+/**
+ *
+ * @this {import('../driver').AndroidDriver}
+ * @param {import('./types').StopScreenRecordingOpts} [options={}]
+ * @returns {Promise<string>}
+ */
+export async function stopRecordingScreen(options = {}) {
+  await verifyScreenRecordIsSupported(this.adb, this.isEmulator());
+
+  if (!_.isEmpty(this._screenRecordingProperties)) {
+    this._screenRecordingProperties.stopped = true;
+  }
+
+  try {
+    await terminateBackgroundScreenRecording(this.adb, false);
+  } catch (err) {
+    this.log.warn(/** @type {Error} */ (err).message);
+    if (!_.isEmpty(this._screenRecordingProperties)) {
+      this.log.warn('The resulting video might be corrupted');
+    }
+  }
+
+  if (_.isEmpty(this._screenRecordingProperties)) {
+    this.log.info(
+      `Screen recording has not been previously started by Appium. There is nothing to stop`
+    );
+    return '';
+  }
+
+  if (
+    this._screenRecordingProperties.recordingProcess &&
+    this._screenRecordingProperties.recordingProcess.isRunning
+  ) {
+    try {
+      await this._screenRecordingProperties.recordingProcess.stop(
+        'SIGINT',
+        PROCESS_SHUTDOWN_TIMEOUT
+      );
+    } catch (e) {
+      this.log.errorAndThrow(
+        `Unable to stop screen recording within ${PROCESS_SHUTDOWN_TIMEOUT}ms`
+      );
+    }
+    this._screenRecordingProperties.recordingProcess = null;
+  }
+
+  if (_.isEmpty(this._screenRecordingProperties.records)) {
+    this.log.errorAndThrow(
+      `No screen recordings have been stored on the device so far. ` +
+        `Are you sure the ${SCREENRECORD_BINARY} utility works as expected?`
+    );
+  }
+
+  const tmpRoot = await tempDir.openDir();
+  try {
+    const localRecords = [];
+    for (const pathOnDevice of this._screenRecordingProperties.records) {
+      const relativePath = path.resolve(tmpRoot, path.posix.basename(pathOnDevice));
+      localRecords.push(relativePath);
+      await this.adb.pull(pathOnDevice, relativePath);
+      await this.adb.rimraf(pathOnDevice);
+    }
+    let resultFilePath = /** @type {string} */ (_.last(localRecords));
+    if (localRecords.length > 1) {
+      this.log.info(`Got ${localRecords.length} screen recordings. Trying to merge them`);
+      try {
+        resultFilePath = await mergeScreenRecords(localRecords, this.log);
+      } catch (e) {
+        this.log.warn(
+          `Cannot merge the recorded files. The most recent screen recording is going to be returned as the result. ` +
+            `Original error: ${/** @type {Error} */ (e).message}`
+        );
+      }
+    }
+    if (_.isEmpty(options.remotePath)) {
+      const {size} = await fs.stat(resultFilePath);
+      this.log.debug(
+        `The size of the resulting screen recording is ${util.toReadableSizeString(size)}`
+      );
+    }
+    return await uploadRecordedMedia(resultFilePath, options.remotePath, options);
+  } finally {
+    await fs.rimraf(tmpRoot);
+    this._screenRecordingProperties = undefined;
+  }
+}
+
+// #region Internal helpers
+
+/**
+ *
  * @param {string} localFile
  * @param {string} [remotePath]
  * @param {import('./types').StopScreenRecordingOpts} uploadOptions
@@ -196,152 +345,7 @@ async function terminateBackgroundScreenRecording(adb, force = true) {
   }
 }
 
-/**
- *
- * @this {import('../driver').AndroidDriver}
- * @param {import('./types').StartScreenRecordingOpts} [options={}]
- * @returns {Promise<string>}
- */
-export async function startRecordingScreen(options = {}) {
-  await verifyScreenRecordIsSupported(this.adb, this.isEmulator());
-
-  let result = '';
-  const {
-    videoSize,
-    timeLimit = DEFAULT_RECORDING_TIME_SEC,
-    bugReport,
-    bitRate,
-    forceRestart,
-  } = options;
-  if (!forceRestart) {
-    result = await this.stopRecordingScreen(options);
-  }
-
-  if (await terminateBackgroundScreenRecording(this.adb, true)) {
-    this.log.warn(
-      `There were some ${SCREENRECORD_BINARY} process leftovers running ` +
-        `in the background. Make sure you stop screen recording each time after it is started, ` +
-        `otherwise the recorded media might quickly exceed all the free space on the device under test.`
-    );
-  }
-
-  if (!_.isEmpty(this._screenRecordingProperties)) {
-    // XXX: this doesn't need to be done in serial, does it?
-    for (const record of this._screenRecordingProperties.records || []) {
-      await this.adb.rimraf(record);
-    }
-    this._screenRecordingProperties = undefined;
-  }
-
-  const timeout = parseFloat(String(timeLimit));
-  if (isNaN(timeout) || timeout > MAX_TIME_SEC || timeout <= 0) {
-    throw new Error(
-      `The timeLimit value must be in range [1, ${MAX_TIME_SEC}] seconds. ` +
-        `The value of '${timeLimit}' has been passed instead.`
-    );
-  }
-
-  this._screenRecordingProperties = {
-    timer: new timing.Timer().start(),
-    videoSize,
-    timeLimit,
-    currentTimeLimit: timeLimit,
-    bitRate,
-    bugReport,
-    records: [],
-    recordingProcess: null,
-    stopped: false,
-  };
-  await scheduleScreenRecord(this.adb, this._screenRecordingProperties, this.log);
-  return result;
-}
-
-/**
- *
- * @this {import('../driver').AndroidDriver}
- * @param {import('./types').StopScreenRecordingOpts} [options={}]
- * @returns {Promise<string>}
- */
-export async function stopRecordingScreen(options = {}) {
-  await verifyScreenRecordIsSupported(this.adb, this.isEmulator());
-
-  if (!_.isEmpty(this._screenRecordingProperties)) {
-    this._screenRecordingProperties.stopped = true;
-  }
-
-  try {
-    await terminateBackgroundScreenRecording(this.adb, false);
-  } catch (err) {
-    this.log.warn(/** @type {Error} */ (err).message);
-    if (!_.isEmpty(this._screenRecordingProperties)) {
-      this.log.warn('The resulting video might be corrupted');
-    }
-  }
-
-  if (_.isEmpty(this._screenRecordingProperties)) {
-    this.log.info(
-      `Screen recording has not been previously started by Appium. There is nothing to stop`
-    );
-    return '';
-  }
-
-  if (
-    this._screenRecordingProperties.recordingProcess &&
-    this._screenRecordingProperties.recordingProcess.isRunning
-  ) {
-    try {
-      await this._screenRecordingProperties.recordingProcess.stop(
-        'SIGINT',
-        PROCESS_SHUTDOWN_TIMEOUT
-      );
-    } catch (e) {
-      this.log.errorAndThrow(
-        `Unable to stop screen recording within ${PROCESS_SHUTDOWN_TIMEOUT}ms`
-      );
-    }
-    this._screenRecordingProperties.recordingProcess = null;
-  }
-
-  if (_.isEmpty(this._screenRecordingProperties.records)) {
-    this.log.errorAndThrow(
-      `No screen recordings have been stored on the device so far. ` +
-        `Are you sure the ${SCREENRECORD_BINARY} utility works as expected?`
-    );
-  }
-
-  const tmpRoot = await tempDir.openDir();
-  try {
-    const localRecords = [];
-    for (const pathOnDevice of this._screenRecordingProperties.records) {
-      const relativePath = path.resolve(tmpRoot, path.posix.basename(pathOnDevice));
-      localRecords.push(relativePath);
-      await this.adb.pull(pathOnDevice, relativePath);
-      await this.adb.rimraf(pathOnDevice);
-    }
-    let resultFilePath = /** @type {string} */ (_.last(localRecords));
-    if (localRecords.length > 1) {
-      this.log.info(`Got ${localRecords.length} screen recordings. Trying to merge them`);
-      try {
-        resultFilePath = await mergeScreenRecords(localRecords, this.log);
-      } catch (e) {
-        this.log.warn(
-          `Cannot merge the recorded files. The most recent screen recording is going to be returned as the result. ` +
-            `Original error: ${/** @type {Error} */ (e).message}`
-        );
-      }
-    }
-    if (_.isEmpty(options.remotePath)) {
-      const {size} = await fs.stat(resultFilePath);
-      this.log.debug(
-        `The size of the resulting screen recording is ${util.toReadableSizeString(size)}`
-      );
-    }
-    return await uploadRecordedMedia(resultFilePath, options.remotePath, options);
-  } finally {
-    await fs.rimraf(tmpRoot);
-    this._screenRecordingProperties = undefined;
-  }
-}
+// #endregion
 
 /**
  * @typedef {import('appium-adb').ADB} ADB
