+++ lib/commands/resources.js
@@ -4,47 +4,6 @@ import {fs, util} from '@appium/support';
 
 /**
  * @this {import('../driver').AndroidDriver}
- * @param {string?} [language]
- * @param {import('../driver').AndroidDriverOpts?} [opts=null]
- * @returns {Promise<string>};
- */
-export async function extractStringsFromResources(language, opts = null) {
-  const caps = opts ?? this.opts;
-
-  /** @type {string|undefined} */
-  let app;
-  try {
-    app = caps.app || (caps.appPackage && caps.tmpDir && await this.adb.pullApk(caps.appPackage, caps.tmpDir));
-  } catch (err) {
-    throw new Error(
-      `Failed to pull an apk from '${caps.appPackage}' to '${caps.tmpDir}'. Original error: ${
-        err.message
-      }`
-    );
-  }
-
-  if (!app || !(await fs.exists(app))) {
-    throw new Error(`Could not extract app strings, no app or package specified`);
-  }
-
-  const stringsTmpDir = path.resolve(String(caps.tmpDir), util.uuidV4());
-  try {
-    this.log.debug(
-      `Extracting strings from '${app}' for the language '${language || 'default'}' into '${stringsTmpDir}'`
-    );
-    const {localPath} = await this.adb.extractStringsFromApk(
-      app,
-      language ?? null,
-      stringsTmpDir
-    );
-    return localPath;
-  } catch (err) {
-    throw new Error(`Could not extract app strings. Original error: ${err.message}`);
-  }
-}
-
-/**
- * @this {import('../driver').AndroidDriver}
  * @param {string?} [language=null]
  * @returns {Promise<import('@appium/types').StringRecord>}}
  */
@@ -90,6 +49,51 @@ export async function ensureDeviceLocale(language, country, script) {
     throw new Error(`Failed to set ${message}`);
   }
 }
+
+// #region Internal helpers
+
+/**
+ * @this {import('../driver').AndroidDriver}
+ * @param {string?} [language]
+ * @param {import('../driver').AndroidDriverOpts?} [opts=null]
+ * @returns {Promise<string>};
+ */
+export async function extractStringsFromResources(language, opts = null) {
+  const caps = opts ?? this.opts;
+
+  /** @type {string|undefined} */
+  let app;
+  try {
+    app = caps.app || (caps.appPackage && caps.tmpDir && await this.adb.pullApk(caps.appPackage, caps.tmpDir));
+  } catch (err) {
+    throw new Error(
+      `Failed to pull an apk from '${caps.appPackage}' to '${caps.tmpDir}'. Original error: ${
+        err.message
+      }`
+    );
+  }
+
+  if (!app || !(await fs.exists(app))) {
+    throw new Error(`Could not extract app strings, no app or package specified`);
+  }
+
+  const stringsTmpDir = path.resolve(String(caps.tmpDir), util.uuidV4());
+  try {
+    this.log.debug(
+      `Extracting strings from '${app}' for the language '${language || 'default'}' into '${stringsTmpDir}'`
+    );
+    const {localPath} = await this.adb.extractStringsFromApk(
+      app,
+      language ?? null,
+      stringsTmpDir
+    );
+    return localPath;
+  } catch (err) {
+    throw new Error(`Could not extract app strings. Original error: ${err.message}`);
+  }
+}
+
+// #endregion
 
 /**
  * @typedef {import('appium-adb').ADB} ADB
